typecast ULL's to uint64_t.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@6449 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ nav/nav_view/src/nav_view/nav_view_panel.cpp
@@ -429,7 +429,7 @@ void NavViewPanel::updateRadiusPosition()
 {
   try
   {
-    tf::Stamped<tf::Pose> robot_pose(btTransform(btQuaternion(0,0,0), btVector3(0,0,0)), ros::Time(0ULL), "base");
+    tf::Stamped<tf::Pose> robot_pose(btTransform(btQuaternion(0,0,0), btVector3(0,0,0)), ros::Time((uint64_t)0ULL), "base");
     tf::Stamped<tf::Pose> map_pose;
 
     tf_client_->transformPose("map", robot_pose, map_pose);
