+++ util/tf/src/tf.cpp
@@ -386,7 +386,7 @@ std::string Transformer::allFramesAsString()
   {
     unsigned int parent_id;
     try{
-      getFrame(counter)->getData(0ULL, temp);
+      getFrame(counter)->getData((uint64_t)0ULL, temp);
       parent_id = temp.parent_frame_id;
     }
     catch (tf::LookupException& ex)
@@ -411,7 +411,7 @@ void Transformer::getFrameStrings(std::vector<std::string> & vec)
   for (unsigned int counter = 1; counter < frames_.size(); counter ++)
   {
     try{
-      getFrame(counter)->getData(0ULL, temp);
+      getFrame(counter)->getData((uint64_t)0ULL, temp);
     }
     catch (tf::LookupException& ex)
     {
