+++ util/tf/test/testBroadcaster.cpp
@@ -49,7 +49,7 @@ public:
         << 0 << 0 << 1 << 3
         << 0 << 0 << 0 << 1;
 
-    broadcaster.sendTransform(btTransform(btQuaternion(0,0,0), btVector3(1,2,3)), 1000000000ULL, "frame1", "frame2");
+    broadcaster.sendTransform(btTransform(btQuaternion(0,0,0), btVector3(1,2,3)), (uint64_t)1000000000ULL, "frame1", "frame2");
     /*    pTFServer->sendEuler("count","count++",1,1,1,1,1,1,ros::Time(100000,100000));
     pTFServer->sendInverseEuler("count","count++",1,1,1,1,1,1,ros::Time(100000,100000));
     pTFServer->sendDH("count","count++",1,1,1,1,ros::Time(100000,100000));
